src/test/java/com/blogspot/ostas/leetcode/all/medium/maximum_gap/SolutionTest.java
package com.blogspot.ostas.leetcode.all.medium.maximum_gap;

import org.junit.jupiter.api.Test;

import static org.assertj.core.api.Assertions.assertThat;

class SolutionTest {

    /*
      Example 1:
    Input: nums = [3,6,9,1]
    Output: 3
    Explanation: The sorted form of the array is [1,3,6,9], either (3,6) or (6,9) has the maximum difference 3.
    Example 2:
    Input: nums = [10]
    Output: 0
    Explanation: The array contains less than 2 elements, therefore return 0.
      Constraints:
    1 <= nums.length <= 105
    0 <= nums[i] <= 109
    */
    @Test
    void example_0() {
        var solution = new Solution();
        int[] nums = new int[]{3, 6, 9, 1};
        int expected = 3;
        var result = solution.maximumGap(nums);
        assertThat(result).isEqualTo(expected);
    }

    @Test
    void example_1() {
        var solution = new Solution();
        int[] nums = new int[]{10};
        int expected = 0;
        var result = solution.maximumGap(nums);
        assertThat(result).isEqualTo(expected);
    }

}
